+++ @types/iframe_client/util.d.ts
@@ -6,7 +6,7 @@
 function getIframeName(): string;
 
 /**
- * 获取使用此函数的脚本在脚本库中的 id, **只能对脚本库使用**
+ * 获取脚本的脚本库 id, **只能在脚本内使用**
  *
  * @returns 脚本库的 id
  */